Consolidated test fixtures
Moved test fixtures inside the test submodule and removed code
duplication.
