+++ test/TestIncludeMissFileWithRole.py
@@ -1,9 +1,6 @@
-import os
 from collections import namedtuple
 
 import pytest
-
-from ansiblelint.runner import Runner
 
 PlayFile = namedtuple('PlayFile', ['name', 'content'])
 
@@ -50,28 +47,6 @@ PLAY_INCLUDED_RELATIVE = PlayFile('roles/include_relative/tasks/included_file.ym
 - debug:
     msg: 'was found & included'
 ''')
-
-
-@pytest.fixture
-def play_file_path(tmp_path):
-    p = tmp_path / 'playbook.yml'
-    return str(p)
-
-
-@pytest.fixture
-def runner(play_file_path, default_rules_collection):
-    return Runner(default_rules_collection, play_file_path, [], [], [])
-
-
-@pytest.fixture
-def _play_files(tmp_path, request):
-    if request.param is None:
-        return
-    for play_file in request.param:
-        print(play_file.name)
-        p = tmp_path / play_file.name
-        os.makedirs(os.path.dirname(p), exist_ok=True)
-        p.write_text(play_file.content)
 
 
 @pytest.mark.parametrize(
